+++ Pharamcy.Domain/Models/MedicineTracking.cs
@@ -4,8 +4,8 @@
     {
         public string? BarCode { get; set; }
         public int Amount { get; set; }
-        public decimal PurchasePrice { get; set; }
-        public decimal SalePrice { get; set; }
+        public double PurchasePrice { get; set; }
+        public double SalePrice { get; set; }
         public int MedicineId { get; set; }
         public virtual Medicine Medicine { get; set; }
     }
